+++ Libplanet.Tests/Net/SwarmTest.cs
@@ -56,12 +56,15 @@ namespace Libplanet.Tests.Net
             {
                 new Swarm(
                     new PrivateKey(),
+                    1,
                     ipAddress: IPAddress.Loopback),
                 new Swarm(
                     new PrivateKey(),
+                    1,
                     ipAddress: IPAddress.Loopback),
                 new Swarm(
                     new PrivateKey(),
+                    1,
                     ipAddress: IPAddress.Loopback),
             };
         }
@@ -306,21 +309,72 @@ namespace Libplanet.Tests.Net
         }
 
         [Fact]
+        public async Task DetectAppProtocolVersion()
+        {
+            var a = new Swarm(
+                new PrivateKey(),
+                ipAddress: IPAddress.Loopback,
+                appProtocolVersion: 2);
+            var b = new Swarm(
+                new PrivateKey(),
+                ipAddress: IPAddress.Loopback,
+                appProtocolVersion: 3);
+
+            var c = new Swarm(
+                new PrivateKey(),
+                ipAddress: IPAddress.Loopback,
+                appProtocolVersion: 2);
+            var d = new Swarm(
+                new PrivateKey(),
+                ipAddress: IPAddress.Loopback,
+                appProtocolVersion: 3);
+
+            BlockChain<DumbAction> chain = _blockchains[0];
+
+            try
+            {
+                await StartAsync(a, chain);
+                await StartAsync(b, chain);
+                await StartAsync(c, chain);
+                await StartAsync(d, chain);
+
+                var peers = new[] { c.AsPeer, d.AsPeer };
+
+                foreach (var peer in peers)
+                {
+                    a.Add(peer);
+                    b.Add(peer);
+                }
+
+                Assert.Equal(new[] { c.AsPeer }, a.ToArray());
+                Assert.Equal(new[] { d.AsPeer }, b.ToArray());
+            }
+            finally
+            {
+                await a.StopAsync();
+                await b.StopAsync();
+            }
+        }
+
+        [Fact]
         public void BeComparedProperly()
         {
             var pk1 = new PrivateKey();
             var pk2 = new PrivateKey();
             var a = new Swarm(
                 pk1,
+                1,
                 ipAddress: IPAddress.Parse("0.0.0.0"),
                 listenPort: 5555);
             var b = new Swarm(
                 pk1,
+                1,
                 ipAddress: IPAddress.Parse("0.0.0.0"),
                 listenPort: 5555,
                 createdAt: a.LastDistributed);
             var c = new Swarm(
                 pk2,
+                1,
                 ipAddress: IPAddress.Parse("0.0.0.0"),
                 listenPort: 5555);
 
@@ -577,7 +631,7 @@ namespace Libplanet.Tests.Net
         {
             Assert.Throws<ArgumentNullException>(() =>
             {
-                new Swarm(null);
+                new Swarm(null, 1);
             });
         }
 
@@ -587,6 +641,7 @@ namespace Libplanet.Tests.Net
             var expected = new IPEndPoint(IPAddress.Parse("1.2.3.4"), 5678);
             Swarm s = new Swarm(
                 new PrivateKey(),
+                1,
                 ipAddress: IPAddress.Parse("1.2.3.4"),
                 listenPort: 5678);
 
@@ -611,7 +666,7 @@ namespace Libplanet.Tests.Net
         public async Task AsPeerThrowSwarmExceptionWhenUnbound()
         {
             Swarm swarm =
-                new Swarm(new PrivateKey(), ipAddress: IPAddress.Loopback);
+                new Swarm(new PrivateKey(), 1, ipAddress: IPAddress.Loopback);
             Assert.Throws<SwarmException>(() => swarm.AsPeer);
 
             await StartAsync(swarm, _blockchains[0]);
@@ -635,9 +690,10 @@ namespace Libplanet.Tests.Net
 
             var seed = new Swarm(
                 new PrivateKey(),
+                1,
                 ipAddress: IPAddress.Loopback);
-            var swarmA = new Swarm(new PrivateKey(), iceServers: iceServers);
-            var swarmB = new Swarm(new PrivateKey(), iceServers: iceServers);
+            var swarmA = new Swarm(new PrivateKey(), 1, iceServers: iceServers);
+            var swarmB = new Swarm(new PrivateKey(), 1, iceServers: iceServers);
 
             try
             {